kernel/rina_ip_dev:rina_ip_dev to dedicated files
Move code for rina_ip_dev netdevice to dedicated files. Also refine setup method
with some properties from TUN interfaces.
